refactor: shorten name
